refactor: slight re-org of types
